// src/modules/dashboard/infrastructure/generated/models/accountDashboardResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AccountDashboardResponseAccountStatus } from './accountDashboardResponseAccountStatus';
import type { AccountOwnerDashboardResponse } from './accountOwnerDashboardResponse';

export interface AccountDashboardResponse {
  id: number;
  bankName: string;
  name: string;
  balance: number;
  currency: string;
  createdAt: string;
  closedAt?: string;
  accountStatus: AccountDashboardResponseAccountStatus;
  description?: string;
  type: string;
  accountOwners: AccountOwnerDashboardResponse[];
}