// src/modules/dashboard/infrastructure/generated/models/dashboardResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AccountDashboardResponse } from './accountDashboardResponse';

export interface DashboardResponse {
  totalAccounts: number;
  totalBalance: number;
  totalActiveAccounts: number;
  accounts: AccountDashboardResponse[];
}